+++ CHANGELOG.md
@@ -7,7 +7,48 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [4.0.0] - X-X-X
+
+## 4.0.8 - 2018-06-04
+
+### Fixed
+
+- Google Maps examples were not working on iOS (https://github.com/CartoDB/carto.js/issues/1995)[https://github.com/CartoDB/carto.js/issues/1995]
+
+## 4.0.7 - 2018-06-04
+
+### Fixes
+- Remove upper limit on Google Maps dependency.
+
+## 4.0.6 - 2018-05-11
+
+### Fixes
+- Fix remove layers
+
+## 4.0.5 - 2018-05-10
+
+- Internal fixes
+
+## 4.0.4 - 2018-05-09
+
+- Internal fixes
+
+## 4.0.3 - 2018-05-09
+
+- Internal fixes
+
+## 4.0.3 - 2018-05-04
+
+- Update zera version
+
+## 4.0.2 - 2018-04-27
+
+- Internal fixes
+
+## 4.0.1 - 2018-04-25
+
+- Upgrading carto.js to gmaps v3.31
+
+## [4.0.0] 2018-04-17
 
 First public release of CARTO.js library